Fixed assemblies

+++ src/domain/usecases/item/DeleteItemUseCase.ts
@@ -1,5 +1,6 @@
-import {mergeMap, Observable} from 'rxjs';
+import {mergeMap, Observable, throwError} from 'rxjs';
 import {inject, injectable} from 'tsyringe';
+import {AuthenticationRepositoryType} from '../../../data/interfaces/repositories/AuthenticationRepositoryType';
 import {ItemRepositoryType} from '../../../data/interfaces/repositories/ItemRepositoryType';
 import {UserRepositoryType} from '../../../data/interfaces/repositories/UserRepositoryType';
 import {DeleteItemUseCaseType} from '../../interfaces/usecases/item/DeleteItemUseCaseType';
@@ -11,9 +12,17 @@ export class DeleteItemUseCase implements DeleteItemUseCaseType {
     private itemRepository: ItemRepositoryType,
     @inject('UserRepositoryType')
     private userRepository: UserRepositoryType,
+    @inject('AuthenticationRepositoryType')
+    private authenticationRepository: AuthenticationRepositoryType,
   ) {}
 
-  deleteItem(userID: string, itemID: string): Observable<void> {
+  deleteItem(itemID: string): Observable<void> {
+    const userID = this.authenticationRepository.authenticatedUserID;
+
+    if (!userID) {
+      return throwError(() => 'There is no userID');
+    }
+
     return this.itemRepository
       .deleteItem(itemID)
       .pipe(
